Moved getAdpVersionStatus to adp-version-service
