+++ packages/data/src/adp-version-data.ts
@@ -39,13 +39,3 @@ export const fetchRemoteAdpVersion = async (): Promise<string> => {
 		)
 	).data.sha;
 };
-
-export const getAdpVersionStatus = async (): Promise<AdpVersionStatus> => {
-	const localAdpVersion = await findLocalAdpVersion();
-	const remoteAdpVersion = await fetchRemoteAdpVersion();
-	
-	return {
-		upToDate: localAdpVersion === remoteAdpVersion,
-		latestAdpVersion: remoteAdpVersion,
-	};
-};
